refactor(matrix): move recovery key code into submodule
Owning abstraction: Matrix recovery-key lifecycle is owned by channels::matrix::recovery, while src/channels/matrix.rs keeps the runtime orchestration and public crate reexports.

Invariant: recovery-key bootstrap, restore, rotate, cleanup-journal handling, and CLI-visible helper paths preserve the same call graph and wire/file shapes after the move.

Validation: /Users/zera/build/carapace/scripts/cargo-serial fmt --check; /Users/zera/build/carapace/scripts/cargo-serial check --all-targets; /Users/zera/build/carapace/scripts/cargo-serial nextest run recovery_key.
